Builds lemmy_routes in parallel with lemmy_apub
