+++ crates/routes/Cargo.toml
@@ -17,7 +17,6 @@ lemmy_db_views = { workspace = true }
 lemmy_db_views_actor = { workspace = true }
 lemmy_db_schema = { workspace = true }
 lemmy_api_common = { workspace = true }
-lemmy_apub = { workspace = true }
 diesel = { workspace = true }
 actix-web = { workspace = true }
 anyhow = { workspace = true }
