+++ crates/routes/src/webfinger.rs
@@ -1,12 +1,11 @@
 use actix_web::{web, web::Query, HttpResponse};
 use anyhow::Context;
 use lemmy_api_common::LemmyContext;
-use lemmy_apub::fetcher::webfinger::{WebfingerLink, WebfingerResponse};
 use lemmy_db_schema::{
   source::{community::Community, person::Person},
   traits::ApubActor,
 };
-use lemmy_utils::{error::LemmyError, location_info};
+use lemmy_utils::{error::LemmyError, location_info, WebfingerLink, WebfingerResponse};
 use serde::Deserialize;
 use url::Url;
