+++ crates/utils/src/lib.rs
@@ -16,7 +16,9 @@ mod test;
 pub mod utils;
 pub mod version;
 
+use serde::{Deserialize, Serialize};
 use std::{fmt, time::Duration};
+use url::Url;
 
 pub type ConnectionId = usize;
 
@@ -29,6 +31,20 @@ impl fmt::Display for IpAddr {
   fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
     write!(f, "{}", self.0)
   }
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct WebfingerLink {
+  pub rel: Option<String>,
+  #[serde(rename = "type")]
+  pub kind: Option<String>,
+  pub href: Option<Url>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct WebfingerResponse {
+  pub subject: String,
+  pub links: Vec<WebfingerLink>,
 }
 
 #[macro_export]
